improve import ergonomics. Improved readme
