+++ src/pyroads/reshape/_functions/widen_by_lane.py
@@ -1,6 +1,6 @@
 from deprecated import deprecated
 
-from ._util.gcd_list import gcd_list
+from ._gcd_list import gcd_list
 from .stretch import stretch
